Fixes #244 - adding compaction settings info. (#261)
* Fixes #244 - adding compaction settings info.

We sent compaction info when there are compactions running. This is fine
from perspective of single and ongoing compactions. But we are configuring
whole status module to have period and on that period it should send some
measurement (health of cluster, unreacgable nodes, state of threads).

I have added compaction settings which are sent on configurable time period
always and single compaction info is sent per compaction when there are
running of compactions.

* Fixing test.